refactor: extract android sdk discovery helpers
